functors In C++ STL

+++ 03-stl-and-memory/2-iterators-algorithms/Function Objects (Functors).cpp
@@ -70,7 +70,7 @@ int main() {
     // -------------------------------
     cout << "Vector before sorting: ";
      
-    for (int i=0; i<v.size(); i++){
+    for (size_t i=0; i<v.size(); i++){
 
      cout<<v.at(i)<<" ";
    }
